requirements.txt
# Voice AI Agent - Dependencies
# Install with: pip install -r requirements.txt

# LangChain framework
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# LLM API
openai>=1.0.0

# Speech-to-Text (faster-whisper)
faster-whisper>=1.0.0

# Text-to-Speech
pyttsx3>=2.90

# Audio recording
sounddevice>=0.4.6
scipy>=1.10.0
numpy>=1.24.0

# Environment variables
python-dotenv>=1.0.0

# Web Interface
streamlit>=1.30.0
